backend/reviews_service/requirements.txt
django>=4.2,<5
djangorestframework
drf-spectacular
psycopg2-binary
gunicorn
